refactor: split component headers
