+++ Engine/include/sge.h
@@ -42,6 +42,22 @@
 #include "runtime/Entity.h"
 #include "geometry/ShapeFactory.h"
 #include "component/Component.h"
+#include "component/TagComponent.h"
+#include "component/SkyboxComponent.h"
+#include "component/RenderableComponent.h"
+#include "component/NativeScriptComponent.h"
+#include "component/PhysicsComponent.h"
+#include "component/CameraComponent.h"
+#include "component/MeshComponent.h"
+#include "component/MaterialComponent.h"
+#include "component/ObjectComponent.h"
+#include "component/ShaderComponent.h"
+#include "component/InstanceBatch.h"
+#include "component/ImageComponent.h"
+#include "component/PlayerController.h"
+#include "component/VolumeComponent.h"
+#include "component/CharacterController.h"
+#include "component/TestComp.h"
 #include "core/Random.h"
 #include "systems/TimeManager.h"
 #include "systems/CommonTextures.h"
