+++ Engine/src/component/RenderableComponent.h
@@ -1,10 +1,24 @@
-#pragma once
+#include "component/Component.h"
+#include "component/ComponentSerializer.h"
 
-//class Renderable
-//{
-//public:
-//	Renderable() = default;
-//
-//private:
-//	int a;
-//};
+struct EngineAPI RenderableComponent : public Component
+{
+RenderableComponent() = default;
+
+static void attachToEntity(std::shared_ptr<Component>, Entity, Scene&);
+
+template <class Archive>
+void serialize(Archive& archive) {
+SERIALIZED_MEMBER(renderTechnique);
+}
+
+enum class RenderTechnique : int
+{
+Forward,
+Deferred
+};
+
+RenderTechnique renderTechnique = RenderTechnique::Deferred;
+};
+
+REGISTER_COMPONENT(RenderableComponent)
