+++ Engine/src/render/DeferredRenderer.cpp
@@ -5,7 +5,10 @@
 #include "core/Engine.h"
 #include "core/Window.h"
 #include "core/Logger.h"
-#include "component/Component.h"
+#include "component/RenderableComponent.h"
+#include "component/ObjectComponent.h"
+#include "component/MeshComponent.h"
+#include "component/ShaderComponent.h"
 #include "component/Transformation.h"
 #include "render/ScreenQuad.h"
 #include "runtime/Scene.h"
