+++ Engine/src/render/Renderer.cpp
@@ -19,8 +19,10 @@
 #include "component/Transformation.h"
 #include "render/Material.h"
 #include "render/Shader.h"
-#include "component/Component.h"
-#include "component/Transformation.h"
+#include "component/RenderableComponent.h"
+#include "component/ObjectComponent.h"
+#include "component/MeshComponent.h"
+#include "component/ShaderComponent.h"
 #include "render/CommonShaders.h"
 #include "geometry/MeshCollection.h"
 #include "render/Graphics.h"
