+++ Engine/src/runtime/Scene.cpp
@@ -20,7 +20,14 @@
 #include "component/Transformation.h"
 #include "geometry/Mesh.h"
 #include "component/RenderableComponent.h"
-#include "component/Component.h"
+#include "component/ObjectComponent.h"
+#include "component/NativeScriptComponent.h"
+#include "component/CameraComponent.h"
+#include "component/MeshComponent.h"
+#include "component/SkyboxComponent.h"
+#include "component/VolumeComponent.h"
+#include "component/ShaderComponent.h"
+#include "component/ImageComponent.h"
 #include "render/Shader.h"
 #include "render/Material.h"
 #include "component/ScriptableEntity.h"
