+++ Engine/src/serialize/Archiver.h
@@ -14,6 +14,17 @@
 #include "entt/entt.hpp"
 
 #include "component/Component.h"
+#include "component/PhysicsComponent.h"
+#include "component/PlayerController.h"
+#include "component/MeshComponent.h"
+#include "component/RenderableComponent.h"
+#include "component/CameraComponent.h"
+#include "component/NativeScriptComponent.h"
+#include "component/MaterialComponent.h"
+#include "component/ObjectComponent.h"
+#include "component/SkyboxComponent.h"
+#include "component/ImageComponent.h"
+#include "component/ShaderComponent.h"
 #include "component/Transformation.h"
 #include "component/ScriptableEntity.h"
 #include "render/Material.h"
