+++ Engine/src/systems/ObjectPicker.cpp
@@ -16,7 +16,9 @@
 #include "runtime/Entity.h"
 #include "texture/Texture.h"
 #include "render/Shader.h"
-#include "component/Component.h"
+#include "component/CameraComponent.h"
+#include "component/SkyboxComponent.h"
+#include "component/MeshComponent.h"
 #include "render/RenderCommand.h"
 #include "geometry/ShapeFactory.h"
 #include "component/Transformation.h"
