+++ Engine/src/systems/ShadowSystem.cpp
@@ -6,7 +6,8 @@
 #include "core/Logger.h"
 #include "runtime/Scene.h"
 #include "render/IRenderer.h"
-#include "component/Component.h"
+#include "component/RenderableComponent.h"
+#include "component/MeshComponent.h"
 #include "core/Engine.h"
 #include "core/Window.h"
 #include "lights/DirectionalLight.h"
