+++ Engine/src/systems/Skybox.cpp
@@ -10,7 +10,9 @@
 #include "render/Shader.h"
 #include "geometry/ShapeFactory.h"
 #include "render/Material.h"
-#include "component/Component.h"
+#include "component/SkyboxComponent.h"
+#include "component/RenderableComponent.h"
+#include "component/ObjectComponent.h"
 #include "runtime/Context.h"
 #include "memory/Assets.h"
 #include "core/Engine.h"
